% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completePhyloData.R
\name{completePhyloData}
\alias{completePhyloData}
\title{Complete phylogeny and traits data}
\usage{
completePhyloData(phylogeny, dataset, traits, taxonName = "animal")
}
\arguments{
\item{phylogeny}{phylogeny with tip labels naming terminal taxa}

\item{dataset}{dataframe with a column including taxa named "animal" equivalent to tip labels in phylogeny}

\item{traits}{character vector}
}
\value{
List with phylogeny and traits data including taxa with both information a
}
\description{
Complete phylogeny and traits data
}
