% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completePhyloData.R
\name{completePhyloFun}
\alias{completePhyloFun}
\title{Complete phylogeny and traits data}
\usage{
completePhyloFun(phylogeny, dataset, traits)
}
\arguments{
\item{phylogeny}{phylogeny with tip labels naming terminal taxa}

\item{dataset}{dataframe with a column including taxa named "animal" equivalent to tip labels in phylogeny}

\item{traits}{character vector}
}
\value{
List with phylogeny and traits data including taxa with both information a
}
\description{
Complete phylogeny and traits data
}
