% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelations.R
\name{computeCorrelations}
\alias{computeCorrelations}
\title{Compute correlations using bivariate phylogenetic mixed models for a given pair of variables.}
\usage{
computeCorrelations(
  variable1,
  variable2,
  dataset,
  phylogeny,
  model.specifications = NULL,
  forceRun = F
)
}
\arguments{
\item{variable1}{(character) Name of the variable. It must be contained in datset.}

\item{variable2}{(character) Name of the variable. It must be contained in datset.}

\item{dataset}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{phylogeny}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{model.specifications}{(list) Mcmcglmm models specifications. See defineModelsSpecification.}

\item{forceRun}{(logical) If false, models already run are not runned again.}
}
\value{

}
\description{
Compute correlations using bivariate phylogenetic mixed models for a given pair of variables.
}
