% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCovariancePartition.R
\name{computeCovariancePartition}
\alias{computeCovariancePartition}
\title{Covariance partition including (or not) environment}
\usage{
computeCovariancePartition(
  traits,
  environmental.variables = NULL,
  dataset,
  phylogeny,
  model.specifications = NULL,
  force.run = T,
  save = T
)
}
\arguments{
\item{traits}{(character) Name of the trait or list of traits. It  must be contained in the dataset.}

\item{environmental.variables}{(character) Names of the environmental variables They must be contained in dataset.}

\item{dataset}{(data frame) Dataset containing the trait of interest and a column named "animal" describing terminal taxa of phylogeny.}

\item{phylogeny}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{model.specifications}{(list) Mcmcglmm models specifications. See defineModelsSpecification.}

\item{force.run}{(logical) If false, models already run are not runned again.}

\item{save}{(logical) If false, resulta re not saved.}
}
\description{
Covariance partition including (or not) environment
}
