% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePhylogeneticSignal.R
\name{computePhylogeneticSignal}
\alias{computePhylogeneticSignal}
\title{Compute phylogenetic signal using univariate phylogenetic mixed models.}
\usage{
computePhylogeneticSignal(
  variable,
  dataset,
  phylogeny,
  model.specifications = NULL,
  forceRun = T
)
}
\arguments{
\item{variable}{(character) Name of the variable. It must be contained in datset.}

\item{dataset}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{phylogeny}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{model.specifications}{Mcmcglmm models specifications. See defineModelsSpecification.}

\item{forceRun}{(logical) If false, models already run are not runned again.}
}
\value{
List containing univariate model results, diagnostics and phylogenetic signal results summary. Phlogenetic signal using Blomberg's K and Pagel's lambda are also reported (using phytools phylosig() fun)
}
\description{
Compute phylogenetic signal using univariate phylogenetic mixed models.
}
