% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseModels.R
\name{diagnoseModels}
\alias{diagnoseModels}
\title{Diangose mcmcglmm model autocorrelation, convergence and effective sample size.}
\usage{
diagnoseModels(model)
}
\arguments{
\item{model}{Mcmcglmm model}
}
\value{
Model diagnostics summary
}
\description{
Diangose mcmcglmm model autocorrelation, convergence and effective sample size.
}
