% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlotNetworkData.R
\name{getPlotNetworkData}
\alias{getPlotNetworkData}
\title{Get data to plot networks}
\usage{
getPlotNetworkData(
  covariance.results,
  variance.results = NULL,
  variance.type = NULL,
  covariance.type = "Total_coordination",
  only.significant = T
)
}
\arguments{
\item{covariance.results}{(data frame) correlations output of computeCovariancePartition()$covarianceResults}

\item{variance.results}{(data frame) phylogenetic signal as reported by computeVariancePartition()$varianceResults}

\item{variance.type}{(character) variable name of the phylogenetic signal as reported in varianceResults}

\item{covariance.type}{(character) variable name of the correlation as reported in correlations}

\item{only.significant}{(logical) if TRUE, only significant correlations are plotted}
}
\description{
Get data to plot networks
}
