% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeTraits.R
\name{imputeTraits}
\alias{imputeTraits}
\title{Impute traits using the evolutionary (or genetic) correlations.}
\usage{
imputeTraits(
  imputationVariables,
  dataset,
  phylogeny,
  covarianceResults = NULL,
  varianceResults = NULL,
  orderCriterium = "Total_coordination",
  numberOfPhyloCoordinates = NULL,
  predictors = NULL,
  prodNAs = 0,
  IterationsNumber = 10,
  clustersNumber = 2,
  forceRun = T,
  parallelization = T,
  specifications = NULL,
  numberOfImputationRounds = 3,
  outputs.dir = ""
)
}
\arguments{
\item{imputationVariables}{(character) Names of the variables with NA where imputations will be implemented. If more than one, covariation among vimputation variables is considered.}

\item{dataset}{(data frame) dataset containing the variable of interest and a column named taxon describing terminal taxa of phylogeny.}

\item{phylogeny}{(phylo) phylogeny with tip labels contained in dataset$taxon}

\item{varianceResults}{(lis)Results from computeVariancePartition() function.}

\item{orderCriterium}{(character) Name of the correlation to be used as ordering criterium (one of teh columns of correlationsTraitsResults)}

\item{numberOfPhyloCoordinates}{(number) Number of phylogenetic axis to include}

\item{predictors}{(character) Names of the variables without NAs used as predictors.}

\item{prodNAs}{(numeric) Proportion of artificial NAs to be introduced in a given dataset. For evaluation puposes.}

\item{IterationsNumber}{(integer) Number of iterations of the imputation processs. Results are summarized.}

\item{clustersNumber}{(integer) Number of clusters to use in parallelization.}

\item{forceRun}{(logical) If false, previously calculated phylogenetic eigenvectors are used.}

\item{parallelization}{(logical) If false, no parallelization is conducted.}

\item{numberOfImputationRounds}{(integer) Number of imputation rounds to implement.}

\item{correlationsTraitsResults}{(list) Results from computeCovariancePartition() function or NULL to compute it internally.}
}
\description{
Impute traits using the evolutionary (or genetic) correlations.
}
