% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeTrEvo.R
\name{initializeTrEvo}
\alias{initializeTrEvo}
\title{Create and define results and outputs folders structure.}
\usage{
initializeTrEvo(folderName = "simulations")
}
\arguments{
\item{folderName}{(character) Folder name.}
}
\value{

}
\description{
Create and define results and outputs folders structure.
}
