% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanOrMode.R
\name{meanOrMode}
\alias{meanOrMode}
\title{Compute mean if variable is numeric or mode if variable is character or factor.}
\usage{
meanOrMode(x)
}
\arguments{
\item{x}{numeric, character or factor vector.}
}
\value{
mean or mode value
}
\description{
Compute mean if variable is numeric or mode if variable is character or factor.
}
