% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeneticSignalTraits.R
\name{phylogeneticSignalTraits}
\alias{phylogeneticSignalTraits}
\title{Retrieve phylogenetic signal for each one of the traits contained in a vector}
\usage{
phylogeneticSignalTraits(
  VARIABLES,
  PHYLOGENY,
  DATASET,
  MODEL.SPECIFICATIONS = NULL,
  FORCERUN = F
)
}
\arguments{
\item{VARIABLES}{(character) Names of the variables. They must be contained in dataset.}

\item{PHYLOGENY}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{DATASET}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{MODEL.SPECIFICATIONS}{(list) Mcmcglmm models specifications. See defineModelsSpecification.}

\item{FORCERUN}{(logical) If false, models already run are not runned again.}
}
\value{

}
\description{
Retrieve phylogenetic signal for each one of the traits contained in a vector
}
