% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationMatrices.R
\name{plotCorrelationMatrices}
\alias{plotCorrelationMatrices}
\title{Plot correlatin matrices.}
\usage{
plotCorrelationMatrices(correlations, phylogenetic.signal, order_vars = NULL)
}
\arguments{
\item{correlations}{(data frame) correlations output of correlationsTraits$correlations.rslts}

\item{phylogenetic.signal}{(data frame) phylogenetic signal as reported by phylogeneticSignalTraits$phylogenetic.signal.rslts}

\item{order_vars}{(character) order of the variables to plot.}
}
\value{

}
\description{
Plot correlatin matrices.
}
