% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot data on the phylogeny}
\usage{
plotData(
  phylogeny,
  dataset,
  variables,
  colorBy = NULL,
  nameBy = NULL,
  panelSpace = 4,
  FontsizeFactor = 0.1
)
}
\arguments{
\item{phylogeny}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{dataset}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{variables}{(character) Names of the variables. They must be contained in dataset.}

\item{colorBy}{(character) name of the variable to colour the tree by.}

\item{nameBy}{(character) name of the taxonomic variable to name the tree by.}

\item{panelSpace}{(numeric)}

\item{FontsizeFactor}{(numeric)}
}
\value{

}
\description{
Plot data on the phylogeny
}
