% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot variance-covariance networks}
\usage{
plotNetwork(
  covariance.results,
  variance.results = NULL,
  variance.type = NULL,
  covariance.type,
  group.variables = NULL,
  order.variables = NULL,
  edge.label = F,
  layout = "circular",
  only.significant = T,
  not.show.variables = NULL,
  threshold = 0,
  label.size = 0.8,
  node.label = NULL,
  plot.metrics = T,
  networkMetrixTextSize = 1,
  displayDegreeAsNodeSize = T
)
}
\arguments{
\item{covariance.results}{(data frame) correlations output of computeCovariancePartition()$covarianceResults}

\item{variance.results}{(data frame) phylogenetic signal as reported by computeVariancePartition()$varianceResults}

\item{variance.type}{(character) variable name of the phylogenetic signal as reported in varianceResults}

\item{covariance.type}{(character) variable name of the correlation as reported in correlations}

\item{group.variables}{(data frame) character vector as long as the number of variables indicating the variable name in the first column and the group of the variable in the second.}

\item{order.variables}{(character) order of the variables to plot}

\item{edge.label}{(logical) if TRUE, edge labels show correlation coefficients}

\item{layout}{"spring" or "circular"}

\item{only.significant}{(logical) if TRUE, only significant correlations are plotted}

\item{not.show.variables}{(character) vector of variables present in correlations data but exclued from the plot}

\item{threshold}{(numeric) correlations to be shown (e.g., those with a absolute value higher than 0.1)}

\item{label.size}{(numeric) size of the node labels}

\item{node.label}{(data frame) character vector as long as the number of variables indicating the variable name in the first column and the new name, which will be shown in the plot.}

\item{networkMetrixTextSize}{(numeric) Size of the text reporting network metrics}

\item{displayDegreeAsNodeSize}{(logical) If true, node size is proportional to its degree.}
}
\description{
Plot variance-covariance networks
}
