% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkData.R
\name{plotNetworkData}
\alias{plotNetworkData}
\title{Interactive 3D plot for visualizing results}
\usage{
plotNetworkData(
  networkData,
  points3Dargs = list(cex = 5, add = FALSE, colkey = FALSE),
  text3Dargs = list(add = TRUE, cex = 2, colkey = FALSE),
  segments3Dargs = list(col = c("red", "green"), lwd = 5, add = TRUE, colkey = FALSE)
)
}
\arguments{
\item{networkData}{List with the coordinates matrix, the node size vector
and the Edge matrix data, usually as returned by \code{\link{getPlotNetworkData}}}

\item{points3Dargs}{List with arguments for \code{\link[plot3D]{points3D}}}

\item{text3Dargs}{List with arguments for \code{\link[plot3D]{text3D}}}

\item{segments3Dargs}{List with arguments for \code{\link[plot3D]{segments3D}}}
}
\value{
plotly object
}
\description{
Interactive 3D plot for visualizing computation results
}
