% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairwiseDistances.R
\name{plotPairwiseDistances}
\alias{plotPairwiseDistances}
\title{Plot pairwise distances using distanace matrices}
\usage{
plotPairwiseDistances(
  matDistX = NULL,
  matDistY = NULL,
  matDistXLab = "Genetic distance",
  matDistYLab = "Chemical distance"
)
}
\arguments{
\item{matDistX}{(matrix) First distance matrix, distances will be displayed in the X axis.}

\item{matDistY}{(matrix) Second distance matrix, distances will be displayed in the Y axis.}

\item{matDistXLab}{(character) X axis label}

\item{matDistYLab}{(character) Y axis label}
}
\description{
Plot pairwise distances using distanace matrices
}
