% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVcv.R
\name{plotVcv}
\alias{plotVcv}
\title{Plot variance-covariance matrix (phylogenetic signal and relative phylogentic and convergent correlations)}
\usage{
plotVcv(
  correlations,
  phylogenetic.signal,
  order_vars = NULL,
  labels = NULL,
  plotLegend = T
)
}
\arguments{
\item{correlations}{(data frame) correlations output of correlationsTraits$correlations.rslts}

\item{phylogenetic.signal}{(data frame) phylogenetic signal as reported by phylogeneticSignalTraits$phylogenetic.signal.rslts}

\item{order_vars}{(character) order of the variables to plot.}

\item{labels}{(data frame) name of the variables and their corresponding labels to be ploted.}
}
\value{

}
\description{
Plot variance-covariance matrix (phylogenetic signal and relative phylogentic and convergent correlations)
}
