% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVcvScatterplots.R
\name{plotVcvScatterplots}
\alias{plotVcvScatterplots}
\title{Plot scatterplots showing the variance-covariancae related to phylogeny-neutral genetic structure and convergence.}
\usage{
plotVcvScatterplots(
  variable1,
  variable2,
  environmental.variables = NULL,
  dataset,
  model.specifications,
  phylogeny
)
}
\arguments{
\item{variable1}{(character) Name of the first variable. It must be contained in dataset.}

\item{variable2}{(character) Name of the second variable. It must be contained in dataset.}

\item{environmental.variables}{(character) Names of the environmental.variables. They must be contained in dataset.}

\item{dataset}{(data frame) dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{model.specifications}{(list) Mcmcglmm models specifications. See defineModelsSpecification.}

\item{phylogeny}{(phylo) phylogeny with tip labels contained in dataset$animal}
}
\description{
Plot scatterplots showing the variance-covariancae related to phylogeny-neutral genetic structure and convergence.
}
