% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataSet.R
\name{simulateDataSet}
\alias{simulateDataSet}
\title{Simulate a list of phylogenetically correlated and non-phylogenetically correlated traits given a phylogeny}
\usage{
simulateDataSet(
  phylogeny = NULL,
  nObs = 100,
  trait.names = c("G1_trait1", "G1_trait2", "G1_env", "G2_trait1", "G2_trait2", "G2_env"),
  vcvMatrix = matrix(c(1, 0.9, 0.8, 0, 0.1, 0.2, 0.9, 1, 0.8, 0, 0.1, 0.2, 0.8, 0.8, 1,
    0, 0.1, 0.2, 0, 0, 0, 1, 0.9, 0.8, 0.1, 0.1, 0.1, 0.9, 1, 0.8, 0.2, 0.2, 0.2, 0.8,
    0.8, 1), ncol = 6)
)
}
\arguments{
\item{phylogeny}{phylogeny under wich simulate phylogenetically correlated traits}

\item{nObs}{Number of observations to simulate if phylogeny is not provided. If provided, the number of terminal taxa of the phylogeny is used.}

\item{trait.names}{Names of the traits to simulate}

\item{vcvMatrix}{Variance covariance matrix setting the covariance among traits. Ncol have to be the same lenght as trait.names.}
}
\value{
A list with a data frame with phylogenetically and non-phylogenetically correlated traits and the variance-covariance matrix used.
}
\description{
Simulate a list of phylogenetically correlated and non-phylogenetically correlated traits given a phylogeny
}
