% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineModelsSpecifications.R
\name{defineModelsSpecifications}
\alias{defineModelsSpecifications}
\title{Define mcmcglmm model specifications}
\usage{
defineModelsSpecifications(
  number.iterations = 100,
  burning = 10,
  thinning = 2,
  uniresponse.prior = list(R = list(V = 1, nu = 0.002), G = list(G1 = list(V = 1, nu =
    0.002))),
  multiresponse.prior = list(R = list(V = diag(2)/2, nu = 2), G = list(G1 = list(V =
    diag(2)/2, nu = 2))),
  triresponse.prior = list(R = list(V = diag(3)/2, nu = 2), G = list(G1 = list(V =
    diag(3)/2, nu = 2)))
)
}
\arguments{
\item{number.iterations}{(integer) Number of iterations of  Markov Chain Montecarlo.}

\item{burning}{(integer) Number of burning iterations.}

\item{thinning}{(integer) Number of thinning iterations.}

\item{uniresponse.prior}{(list) Parameters for the random effects of uniresponse models.}

\item{multiresponse.prior}{(list) Parameters for the random effects of multiresponse models}
}
\value{
List of mcmcglmm model specifications.
}
\description{
Define mcmcglmm model specifications
}
