% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeTraits.R
\name{imputeTraits}
\alias{imputeTraits}
\title{Impute traits using phylogenetic, environmental and traits relationships}
\usage{
imputeTraits(
  variables_to_impute = NULL,
  dataset = NULL,
  terminal_taxon = NULL,
  phylogeny = NULL,
  correlation_results = NULL,
  variance_results = NULL,
  number_of_phylo_axis = NULL,
  predictors = NULL,
  proportion_NAs = 0,
  number_iterations = 10,
  number_clusters = 2,
  model_specifications = NULL,
  save = F,
  force_run = T
)
}
\arguments{
\item{variables_to_impute}{(\emph{character}). Names of the variables with NAs where imputations will be implemented.
If more than one, covariation among imputation variables is considered to decide whether to use them as predictors or not.}

\item{dataset}{(\emph{data frame}). Data frame containing the variable of interest with missing values and a column describing terminal taxa of phylogeny (e.g., species).}

\item{terminal_taxon}{(\emph{character}). Terminal taxon as named in the dataset (e.g., species).}

\item{phylogeny}{(\emph{phylo}). Phylogeny with tip labels contained in dataset.}

\item{correlation_results}{(\emph{list}). Correlation results from computeVarianceCovariance() function or NULL to compute it internally for the variables to be imputed using the data and phylogeny provided (default).}

\item{variance_results}{(\emph{list}). Variance results from computeVarianceCovariance() function.}

\item{number_of_phylo_axis}{(\emph{integer}). Number of phylogenetic axis to include as predictors.}

\item{predictors}{(\emph{character}). Names of the variables without NAs that will be considered as potential. These varaibles need to be included in the dataset and need to be complete (no NAs).}

\item{proportion_NAs}{(\emph{numeric}). Between 0 and 1. Proportion of artificial NAs to be introduced in a given dataset. For evaluation puposes. Default set to 0, so no extra NAs are produced.}

\item{number_iterations}{(\emph{integer}). Number of iterations of the imputation process. Results reported are summarized as mean and standard deviation for each variable to be imputed.}

\item{number_clusters}{(\emph{integer}). Number of clusters to use in parallelization. If set to one, no paralelization is performed (not recommended). Default is set to 2.}

\item{model_specifications}{(\emph{list}). Mcmcglmm models specifications as specified by the defineModelsSpecification of this package. If not defined,
the  default of the defineModelsSpecification function is used.}

\item{save}{(\emph{logical}) If false, results are not saved in the outputs folder.}

\item{force_run}{(\emph{logical}) If false, previously calculated phylogenetic axis (saved internally) are used. This can be useful when using big phylogenies, as the calculation
of the phylogenetic axis can take some time.}
}
\description{
Predict missing values using phylogeny, environmental and trait data as predictors in random forests when it shows a relationship with the variable
to be imputed. The methodology implements 3 rounds in which the actualized predicted values are used to predict the rest of the variables to impute.
}
\examples{
\dontrun{
# Simulate example data
simulated_traits.data <- simulateDataSet()

# Impute missing values (created within the function)
imputed.data <- imputetraits(
variables_to_impute = c("phylo_G1_trait1", "phylo_G1_trait2"),
dataset = simulated_traits.data$data,
phylogeny = simulated_traits.data$phylogeny,
proportion_NAs = 0.2
)
}
}
