% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeTraits.R
\name{imputeTraits}
\alias{imputeTraits}
\title{Impute traits using the evolutionary (or genetic) correlations.}
\usage{
imputeTraits(
  DATASET,
  PHYLOGENY,
  correlationsTraitsResults,
  imputationVariables,
  PREDICTORS = c(paste0("Phylo_axis_", 1:10)),
  prodNAs = 0,
  IterationsNumber = 10,
  clustersNumber = 2,
  FORCERUN = T
)
}
\arguments{
\item{DATASET}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{PHYLOGENY}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{correlationsTraitsResults}{(list) Results from correlationsTraits() function.}

\item{imputationVariables}{(character) Names of the variables with NA where imputations will be implemented. If more than one, covariation among vimputation variables is considered.}

\item{PREDICTORS}{(character) Names of the variables without NAs used as predictors.}

\item{prodNAs}{(numeric) Proportion of artificial NAs to be introduced in a given dataset. For evaluation puposes.}

\item{IterationsNumber}{(integer) Number of iterations of the imputation processs. Results are summarized.}

\item{clustersNumber}{(integer) Number of clusters to use in parallelization.}

\item{FORCERUN}{(logical) If false, previously calculated phylogenetic eigenvectors are used.}
}
\value{

}
\description{
Impute traits using the evolutionary (or genetic) correlations.
}
