% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanOrMode.R
\name{meanOrMode}
\alias{meanOrMode}
\title{Return mean if variable is numeric or mode if variable is character or factor.}
\usage{
meanOrMode(x)
}
\arguments{
\item{x}{(\emph{numeric}, \emph{character} or \emph{factor}). Vector from which the mean or mode will be computed for.}
}
\value{
(\emph{numeric} for mean or \emph{character} for mode). Mean or mode value.
}
\description{
Return mean if variable is numeric or mode if variable is character or factor.
}
