% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot data on the phylogeny}
\usage{
plotData(
  variables,
  dataset,
  phylogeny,
  terminal.taxon = "species",
  color_by = NULL,
  name_by = NULL,
  panel_space = 4,
  fontsize_factor = 0.1
)
}
\arguments{
\item{variables}{(\emph{character}). Names of the variable or variables to be ploted as they appear in data.}

\item{dataset}{(\emph{data frame}). Data containing the variable of interest and a column naming the terminal taxa as it appears in the phylogeny (e.g., species names).}

\item{phylogeny}{(\emph{phylo}). Phylogeny with tip labels contained in data.}

\item{color_by}{(\emph{character}). Name of the variable to colour the tree by, if desired.}

\item{name_by}{(\emph{character}). Name of the taxonomic variable to display in the phylogeny panel.}

\item{panel_space}{(\emph{numeric}). Size of each one of the panels (one for the phylogeny and one per variable included in the "variables" argument).}

\item{fontsize_factor}{(\emph{numeric}). Font size multiplier. It modifies the size of the font.}
}
\description{
Plot variables as barplots for each tip of the phylogeny with data available. Several variables can be plotted at the same time.
}
\examples{
\dontrun{
# Simulate example data
simulated_traits.data <- simulateDataSet()

# Plot G1_trait1 data on simulated phylogeny

plotData(variables = "G1_trait1",
dataset = simulated_traits.data$data,
phylogeny = simulatedTraits.data$phylogeny,,
terminal.taxon = "species"
)
}
}
