% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot data on the phylogeny}
\usage{
plotData(
  variables,
  dataset,
  phylogeny,
  color.by = NULL,
  nameBy = NULL,
  panel.space = 4,
  Fontsize.factor = 0.1
)
}
\arguments{
\item{variables}{(character) Names of the variables. They must be contained in dataset.}

\item{dataset}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{phylogeny}{(phylo) Phylogeny with tip labels contained in dataset$animal}

\item{color.by}{(character) name of the variable to colour the tree by.}

\item{nameBy}{(character) name of the taxonomic variable to name the tree by.}

\item{panel.space}{(numeric)}

\item{Fontsize.factor}{(numeric)}
}
\description{
Plot data on the phylogeny
}
