% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVcvScatterplots.R
\name{plotVcvScatterplots}
\alias{plotVcvScatterplots}
\title{Plot scatterplots showing the variance-covariancae related to phylogeny-neutral genetic structure and convergence.}
\usage{
plotVcvScatterplots(
  VARIABLE1 = "ln_Ks",
  VARIABLE2 = "ln_negP50",
  PREDICTORS,
  DATASET,
  MODEL.SPECIFICATIONS,
  PHYLOGENY
)
}
\arguments{
\item{VARIABLE1}{(character) Name of the first variable. It must be contained in dataset.}

\item{VARIABLE2}{(character) Name of the second variable. It must be contained in dataset.}

\item{PREDICTORS}{(character) Names of the predictors. They must be contained in dataset.}

\item{DATASET}{(data frame) Dataset containing the variable of interest and a column named animal describing terminal taxa of phylogeny.}

\item{MODEL.SPECIFICATIONS}{(list) Mcmcglmm models specifications. See defineModelsSpecification.}

\item{PHYLOGENY}{(phylo) Phylogeny with tip labels contained in dataset$animal}
}
\value{

}
\description{
Plot scatterplots showing the variance-covariancae related to phylogeny-neutral genetic structure and convergence.
}
