% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataSet.R
\name{simulateDataSet}
\alias{simulateDataSet}
\title{Simulate a list of phylogenetically correlated and non-phylogenetically correlated traits given a phylogeny}
\usage{
simulateDataSet(
  phylogeny = NA,
  nObs = 100,
  traitNames = c("HC_1", "HC_2", "HC_predictor", "LC_1", "LC_2", "LC_predictor"),
  vcvMatrix = matrix(c(1, 0.9, 0.8, 0, 0.1, 0.2, 0.9, 1, 0.8, 0, 0.1, 0.2, 0.8, 0.8, 1,
    0, 0.1, 0.2, 0, 0, 0, 1, 0.9, 0.8, 0.1, 0.1, 0.1, 0.9, 1, 0.8, 0.2, 0.2, 0.2, 0.8,
    0.8, 1), ncol = 6)
)
}
\arguments{
\item{phylogeny}{phylogeny under wich simulate phylogenetically correlated traits}

\item{nObs}{Number of observations to simulate if phylogeny is not provided. If provided, the number of terminal taxa of the phylogeny is used.}

\item{traitNames}{Names of the traits to simulate}

\item{vcvMatrix}{Variance covariance matrix setting the covariance among traits. Ncol have to be the same lenght as traitNames.}
}
\value{
A data frame with phylogenetically and non-phylogenetically correlated traits.
}
\description{
Simulate a list of phylogenetically correlated and non-phylogenetically correlated traits given a phylogeny
}
